PR feedback: extract utility file
