+++ frontend/src/components/sessions/data-visualization/StatisticsSection/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Flex, Grid, Text } from "@chakra-ui/react";
 
-import StatisticCard, { type StatisticCardVariant } from "./StatisticCard";
+import StatisticCard, { type StatisticCardVariant } from "../StatisticCard";
 
 export type StatisticsConfig = {
   [key: string]: {
@@ -18,17 +18,6 @@ type Statistics<T> = {
 
 const getKeys = <T,>(obj: StatisticsConfig) =>
   Object.keys(obj) as (keyof T & string)[];
-
-export const NUMBER_FORMAT = (value: number) => value.toFixed(0);
-export const PERCENTAGE_FORMAT = (value: number) => `${NUMBER_FORMAT(value)}%`;
-export const NTH_FORMAT = (value: number) => {
-  const lastDigit = value % 10;
-  let suffix = "th";
-  if (lastDigit === 1) suffix = "st";
-  if (lastDigit === 2) suffix = "nd";
-  if (lastDigit === 3) suffix = "rd";
-  return `${NUMBER_FORMAT(value)}${suffix}`;
-};
 
 const StatisticsSection = <T,>({
   config,
